chore(components)Seperate components into different files - Create `AmortizationChart.js`and transfer its class from `app.js` - Create `MortgageCalculator.js`and transfer its class from `app.js` - Create `Header.js`and transfer its class from `app.js`
